requirements.txt
langchain
langchain-community
faiss-cpu
tqdm
python-dotenv
ollama
langchain-ollama
pypdf
unstructured
python-docx
pandas 